move variant trim_alt to binf module
